==> notebooks/01 - KNN_first_approach.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing
from sklearn.neighbors import KNeighborsRegressor
import csv

### Primer acercamiento a Machine Learning (KNN)

dfTest = pd.read_csv('../data/TEST_SET/properati_dataset_testing_noprice.csv', parse_dates=['created_on'], infer_datetime_format=True)
dfTrain = pd.read_csv('../data/output_2/output_2.csv', parse_dates=['created_on'], infer_datetime_format=True)

dfTest.info()

dfTrain.info()

___
### Función para corregir el Test Set:
> + Debe traducir al inglés **property_type**
> + Eliminar la columna operation ya que supongo que son todos venta (por más que haya 1 alquiler)
> + Corregir con la descripcion el unico registro de Buenos Aires Interior
> + Debe mantener el orden inicial del set ya que se chequean los resultados por indices de array.

> #### Features restantes:
+ id
+ created_on
+ property_type
+ place_name
+ state_name
+ lat
+ lon
+ surface_total_in_m2

def _translate(row) :
    
    english = ['apartment', 'PH', 'house', 'store']
    spanish = ['departamento', 'ph', 'casa', 'local']
    
    if row.property_type in spanish :
        return english[spanish.index(row.property_type)]
    
    return row.property_type



def correctTestSet(testSet) :
    testSet['order'] = pd.Series(xrange(len(dfTest)))
    testSet.property_type = testSet.apply(_translate, axis=1)
    testSet.iat()[1, 4] = 'Palermo'
    testSet.iat()[1, 7] = 'Capital Federal'
    testSet.iat()[1, 14] = 2
    testSet['year_created'] = pd.Series(testSet.created_on.dt.year)
    testSet['month_created'] = pd.Series(testSet.created_on.dt.month)
    testSet['day_created'] = pd.Series(testSet.created_on.dt.day)
    return testSet.sort_values(by='order').loc[:, ('id', 'year_created', 'month_created', 'day_created', 
                                                   'property_type', 'place_name', 'state_name', 'lat', 'lon', 
                                                   'surface_total_in_m2', 'rooms') ]
    
    

dfTest.loc[dfTest.state_name.str.contains('Inter')].iat()[0, len(dfTest.columns) - 1]

dfTest.loc[dfTest.id == 3633]

dfTest = correctTestSet(dfTest)

dfTest.loc[dfTest.id == 3633]

### TestSet Corregido:

dfTest.head()

___

### Corrigiendo TrainSet:
> + Valores muy bajos o extremadamente altos en **surface_total_in_m2** y **price_usd_per_m2**.
> + Filtrar latitudes y longitudes muy por fuera de Bs. As.
> + Filtrar precios irrealmente bajos y altos.
> + Eliminar columna currency.

def _pointInsideBsAs(lat, lon) :
    NO, NE, SE, SO = (-34.2560, -60.1354), (-34.2560, -57.0902), (-36.0219, -57.0902), (-36.0219, -60.1354)    
    return not ( (lat > NO[0]) or (lon < NO[1]) or (lon > NE[1]) or (lat < SE[0]) )

def pointInsideBsAs(latSeries, lonSeries) :    
    result = []
    
    for lat, lon in zip(latSeries, lonSeries) :
        result.append(_pointInsideBsAs(lat, lon))
    
    return np.array(result, dtype='bool')

def correctTrainSet(trainSet): 
    trainSet = trainSet.loc[ pointInsideBsAs(trainSet.lat, trainSet.lon) ]
    trainSet = trainSet.loc[ (trainSet.price_aprox_usd > 11) & (trainSet.price_aprox_usd < 20000000) ]
    trainSet = trainSet.loc[ (trainSet.surface_total_in_m2 > 11) & (trainSet.surface_total_in_m2 < 15000) ]
    #trainSet = trainSet.loc[ (trainSet.surface_covered_in_m2 > 11) & (trainSet.surface_covered_in_m2 < 15000) ]
    #trainSet = trainSet.loc[ trainSet.rooms < 11 ]
    #trainSet = trainSet.loc[ trainSet.expenses < 3500 ]
    trainSet['year_created'] = pd.Series(trainSet.created_on.dt.year)
    trainSet['month_created'] = pd.Series(trainSet.created_on.dt.month)
    trainSet['day_created'] = pd.Series(trainSet.created_on.dt.day)
    return trainSet.loc[:, ('id', 'year_created', 'month_created', 'day_created', 
                            'property_type', 'place_name', 'state_name', 'lat', 'lon', 
                            'surface_total_in_m2', 
                            'rooms', 'price_aprox_usd') ]

dfTrain = correctTrainSet(dfTrain)

dfTrain.info()

dfTrain.describe()

___
### Corriendo el primer KNN

dfTrain = dfTrain.drop('id', axis=1).rename(columns={'price_aprox_usd':'price_usd'})

dfTrain.head()

dfTest.head()

# Tengo que pasar property_type, place_name y state_name a datos numéricos:

le = preprocessing.LabelEncoder()
le.fit(np.array(list(set(list(dfTrain.property_type.unique()) + list(dfTest.property_type.unique()))), dtype='object'))
dfTrain.property_type = le.transform(dfTrain.property_type)
dfTest.property_type = le.transform(dfTest.property_type)

le = preprocessing.LabelEncoder()
le.fit(np.array(list(set(list(dfTrain.place_name.unique()) + list(dfTest.place_name.unique()))), dtype='object'))
dfTrain.place_name = le.transform(dfTrain.place_name)
dfTest.place_name = le.transform(dfTest.place_name)

le = preprocessing.LabelEncoder()
le.fit(np.array(list(set(list(dfTrain.state_name.unique()) + list(dfTest.state_name.unique()))), dtype='object'))
dfTrain.state_name = le.transform(dfTrain.state_name)
dfTest.state_name = le.transform(dfTest.state_name)

# Lleno los valores faltantes (NaN's) con la clase predominante en esa columna

imp = preprocessing.Imputer(missing_values='NaN', strategy='most_frequent')
imp.fit(dfTrain[['rooms']])
dfTrain.rooms = imp.transform(dfTrain[['rooms']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTrain[['lat']])
dfTrain.lat = imp.transform(dfTrain[['lat']]).ravel()
imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTrain[['lon']])
dfTrain.lon = imp.transform(dfTrain[['lon']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='most_frequent')
imp.fit(dfTest[['rooms']])
dfTest.rooms = imp.transform(dfTest[['rooms']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTest[['lat']])
dfTest.lat = imp.transform(dfTest[['lat']]).ravel()
imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTest[['lon']])
dfTest.lon = imp.transform(dfTest[['lon']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTest[['surface_total_in_m2']])
dfTest.surface_total_in_m2 = imp.transform(dfTest[['surface_total_in_m2']]).ravel()

dfTrain.info()

dfTest.info()

# Genero TRAIN y TARGET
target = dfTrain.price_usd
train = dfTrain.drop('price_usd', axis=1)

testIds = dfTest.loc[:, ('id')]
testVals = dfTest.drop('id', axis=1)

knn = KNeighborsRegressor(n_neighbors=30, weights='distance', 
                          algorithm='kd_tree', leaf_size=30, 
                          metric='euclidean', n_jobs=-1)

knn.fit(train, target)

predictions = knn.predict(testVals)

result = pd.DataFrame({'id':testIds, 'price_usd':predictions})
result.head()

# Escribo a archivo
result.to_csv('../data/predictions/1st_acercamiento.csv', index=False)
