==> notebooks/02 - KNN_2nd_approach.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing
from sklearn.neighbors import KNeighborsRegressor
import geopy.distance

### Primer acercamiento a Machine Learning (KNN)

dfTest = pd.read_csv('../data/TRAIN_TEST_corrected/test_corrected.csv')
dfTrain = pd.read_csv('../data/TRAIN_TEST_corrected/train_corrected.csv')
dfExternalCoords = pd.read_csv('../data/externalData/coordenadas_barrios.csv')

dfTest.info()

dfTrain.info()

dfExternalCoords.info()

# Filtro los 3 registros de los cuales no tengo informacion de location
dfExternalCoords = dfExternalCoords.loc[pd.notnull(dfExternalCoords.lat)]
dfExternalCoords.head()

## ToDo:
+ Resolver NaN's en lat / lon con externalData.
+ Reemplazar place_name con distancia al obelisco.
+ Z-codes Standardization.
+ OneHotEncode property_type y state_name.
+ Make Submit & Comparison
+ Cross Validation K-Folds -> Make Submit & Comparison

dfTrain.head(2)

dfTest.head(2)

___
### Corriendo el primer KNN

dfUtil = dfTrain.loc[:, ('place_name', 'year_created')]\
                .groupby(by='place_name')\
                .count().reset_index()\
                .sort_values(by='year_created', ascending=False)\
                .rename(columns={'year_created':'place_ocurrences'})
print 'Cantidad de barrios con menos de 100 publicaciones:', dfUtil.loc[dfUtil.place_ocurrences < 100, ('place_ocurrences')].sum()

# ONE-TO-K ENCODING: PROPERTY_TYPE
prop_type_labelEncoder = preprocessing.LabelEncoder()
prop_type_labelEncoder.fit(np.array(list(dfTrain.property_type.unique()), dtype='object'))

trainPropTypeIntEncoded = prop_type_labelEncoder.transform(dfTrain.property_type)
testPropTypeIntEncoded = prop_type_labelEncoder.transform(dfTest.property_type)

enc = preprocessing.OneHotEncoder(sparse=False)
matrixTrainPropTypeEnc = enc.fit_transform(trainPropTypeIntEncoded.reshape(len(trainPropTypeIntEncoded), 1))
matrixTestPropTypeEnc = enc.transform(testPropTypeIntEncoded.reshape(len(testPropTypeIntEncoded), 1))

oneHotCodeCategoriesTrain = pd.DataFrame(matrixTrainPropTypeEnc, columns=prop_type_labelEncoder.classes_)
oneHotCodeCategoriesTest = pd.DataFrame(matrixTestPropTypeEnc, columns=prop_type_labelEncoder.classes_)

dfTrain = pd.concat([dfTrain.reset_index(drop=True), oneHotCodeCategoriesTrain.reset_index(drop=True)], 
                     axis=1).drop('property_type', axis=1)

dfTest = pd.concat([dfTest.reset_index(drop=True), oneHotCodeCategoriesTest.reset_index(drop=True)],
                    axis=1).drop('property_type', axis=1)

dfTrain.head(2)

dfTest.head(2)

### Reemplazo de place_name y state_name por distancia al obelisco
> + Si tiene lat y lon la dejo y calculo su distancia al obelisco.
> + Si no tiene lat y lon, las estraigo de los datos externos traidos de Gmaps

# Agrego un entero para poder volver al orden original luego de el procesamiento
dfTrain.loc[:, ('order')] = pd.Series(xrange(len(dfTrain)))
dfTest.loc[:, ('order')] = pd.Series(xrange(len(dfTest)))

# Calculo distancias al obelisco en donde pueda y en las otras dejo NaN :

def distanceToObeliscoMeters(seriesLat, seriesLon) :
    distance_2_obelisco = []
    obelisco = (-34.6037389, -58.3815704)
    
    for lat, lon in zip(seriesLat, seriesLon) :        
        dist_m = np.nan if (pd.isnull(lat) or pd.isnull(lon)) else geopy.distance.vincenty(obelisco, (lat, lon)).m
        distance_2_obelisco.append(dist_m)
        
    return pd.Series(distance_2_obelisco)

distancesObeliscoTrain = distanceToObeliscoMeters(dfTrain.lat, dfTrain.lon)
distancesObeliscoTest = distanceToObeliscoMeters(dfTest.lat, dfTest.lon)
dfTrain.insert(column='distance_to_obelisco_mtrs', loc=4, value=distancesObeliscoTrain)
dfTest.insert(column='distance_to_obelisco_mtrs', loc=4, value=distancesObeliscoTest)

dfUpdateTrain = dfTrain.loc[pd.isnull(dfTrain.distance_to_obelisco_mtrs)].rename(columns={'place_name':'place'})\
                                                                         .merge(dfExternalCoords, how='inner', on=['place'])\
                                                                         .drop(['lat_x', 'lon_x', 'distance_to_obelisco_mtrs_x'], axis=1)\
                                                                         .rename(columns={'lat_y':'lat', 'lon_y':'lon', 'distance_to_obelisco_mtrs_y':'distance_to_obelisco_mtrs'})
            
dfUpdateTest = dfTest.loc[pd.isnull(dfTest.distance_to_obelisco_mtrs)].rename(columns={'place_name':'place'})\
                                                                      .merge(dfExternalCoords, how='inner', on=['place'])\
                                                                      .drop(['lat_x', 'lon_x', 'distance_to_obelisco_mtrs_x'], axis=1)\
                                                                      .rename(columns={'lat_y':'lat', 'lon_y':'lon', 'distance_to_obelisco_mtrs_y':'distance_to_obelisco_mtrs'})

dfTrain.set_index('id', drop=True, inplace=True)
dfTest.set_index('id', drop=True, inplace=True)
dfUpdateTrain.set_index('id', drop=True, inplace=True)
dfUpdateTest.set_index('id', drop=True, inplace=True)
dfTrain.update(dfUpdateTrain)
dfTest.update(dfUpdateTest)

dfTrain = dfTrain.reset_index().sort_values(by='order').drop(['place_name', 'state_name', 'order'], axis=1)
dfTest = dfTest.reset_index().sort_values(by='order').drop(['place_name', 'state_name', 'order'], axis=1)

dfTrain.info()

dfTest.info()

### Lleno los valores faltantes (NaN's) con la clase predominante en esa columna

imp = preprocessing.Imputer(missing_values='NaN', strategy='most_frequent')
imp.fit(dfTrain[['rooms']])
dfTrain.rooms = imp.transform(dfTrain[['rooms']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='most_frequent')
imp.fit(dfTest[['rooms']])
dfTest.rooms = imp.transform(dfTest[['rooms']]).ravel()

imp = preprocessing.Imputer(missing_values='NaN', strategy='mean')
imp.fit(dfTest[['surface_total_in_m2']])
dfTest.surface_total_in_m2 = imp.transform(dfTest[['surface_total_in_m2']]).ravel()

dfTrain.info()

dfTest.info()

dfTrain.head()

dfTest.head()

# Genero TRAIN y TARGET
dfTrain = dfTrain.rename(columns={'price_aprox_usd':'price_usd'}).dropna()
target = dfTrain.price_usd
train = dfTrain.drop(['price_usd', 'id'], axis=1)

testIds = dfTest.loc[:, ('id')]
testVals = dfTest.drop('id', axis=1)

print (len(target), len(train))
print (len(testIds), len(testVals))

knn = KNeighborsRegressor(n_neighbors=30, weights='distance', 
                          algorithm='kd_tree', leaf_size=30, 
                          metric='euclidean', n_jobs=-1)

knn.fit(train, target)

predictions = knn.predict(testVals)

result = pd.DataFrame({'id':testIds, 'price_usd':predictions})
result.head()

result.to_csv('../data/predictions/2nd_acercamiento.csv', index=False)

### Guardo Train y Test para el proximo aproximamiento

dfTrain.head(1)

dfTest.head(1)

# EXPORTO los CSV's listos para la 3er aproximacion

dfTrain.to_csv('../data/TRAIN_TEST_corrected2/train_corrected2.csv', index=False)
dfTest.to_csv('../data/TRAIN_TEST_corrected2/test_corrected2.csv', index=False)
