==> notebooks/03 - KNN-3rd_approach.ipynb <==
import pandas as pd
import numpy as np
from sklearn import preprocessing
from sklearn.neighbors import KNeighborsRegressor

dfTrain = pd.read_csv('../data/TRAIN_TEST_corrected2/train_corrected2.csv')
dfTest = pd.read_csv('../data/TRAIN_TEST_corrected2/test_corrected2.csv')

dfTrain.info()

dfTest.info()

dfTrain.head(3)

dfTest.head(3)

target = dfTrain.price_usd
dfTrain.drop(['price_usd'], axis=1, inplace=True)

train = pd.DataFrame(preprocessing.scale(dfTrain.drop(['id'], axis=1)), 
                     columns=['year_created', 'month_created', 'day_created', 
                              'distance_to_obelisco_mtrs', 'lat', 'lon', 
                              'surface_total_in_m2', 'rooms', 'PH', 
                              'apartment', 'house', 'store'])

testIds = dfTest.id
dfTest.drop(['id'], axis=1, inplace=True)

testVals = pd.DataFrame(preprocessing.scale(dfTest), 
                        columns=['year_created', 'month_created', 'day_created', 
                                 'distance_to_obelisco_mtrs', 'lat', 'lon', 
                                 'surface_total_in_m2', 'rooms', 'PH', 
                                 'apartment', 'house', 'store'])

train.head()

testVals.head()

knn = KNeighborsRegressor(n_neighbors=30, weights='distance', 
                          algorithm='kd_tree', leaf_size=30, 
                          metric='euclidean', n_jobs=-1)

knn.fit(train, target)

predictions = knn.predict(testVals)

result = pd.DataFrame({'id':testIds, 'price_usd':predictions})
result.head()

result.to_csv('../data/predictions/3rd_acercamiento.csv', index=False)
