==> notebooks/07 - SVM_1st_approach.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.svm import LinearSVR

dfTrain = pd.read_csv('../data/TRAIN_TEST_corrected3/train_corrected3.csv')
dfTest = pd.read_csv('../data/TRAIN_TEST_corrected3/test_corrected3.csv')

train = dfTrain.drop(['price_usd', 'id'], axis=1)
target = dfTrain.price_usd

test = dfTest.drop(['id'], axis=1)
test_ids = dfTest.id

# DATA SCALING

scaler = StandardScaler()
train = pd.DataFrame(scaler.fit_transform(train), columns=train.columns)
test = pd.DataFrame(scaler.transform(test), columns=test.columns)

# ENTRENAMIENTO

svr = LinearSVR(C=10)
svr.fit(train, target)

# PREDICCION

predictions = svr.predict(test)
result = pd.DataFrame({'id':test_ids, 'price_usd':predictions})

result.head()

result.describe()

result.to_csv('../data/predictions/LinearSVR_Undated_Scored_Scaled_10C.csv', index=False)
