==> notebooks/12 - BayesianRidge_CrossValidation.ipynb <==
import sys
import time
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.linear_model import BayesianRidge
from matplotlib import pyplot as plt


start = time.time()

dfTrain = pd.read_csv('../data/TRAIN_TEST_corrected3/train_corrected3.csv')
dfTrain.reset_index(drop=True)

train = dfTrain.drop(['id', 'price_usd', 'day_created', 'month_created', 'year_created'], axis=1)
scaler = StandardScaler()
train = pd.DataFrame(scaler.fit_transform(train), columns=train.columns)
target = dfTrain.price_usd

train.head()

param_grid = [
                { 'alpha_1'  : [1/float(10), 1],
                  'alpha_2'  : [1/float(1000000), 1/float(100000), 1/float(10000), 1/float(1000), 1/float(100), 1/float(10), 1],
                  'lambda_2' : [1/float(1000000), 1/float(100000), 1/float(10000), 1/float(1000), 1/float(100), 1/float(10), 1] }
              ]

scoring = {'MSE': 'neg_mean_squared_error'}

gs = GridSearchCV(BayesianRidge(), scoring=scoring, 
                  param_grid=param_grid, cv=10, refit='MSE', 
                  return_train_score=False, n_jobs=-1)

gs.fit(train, target);
results = gs.cv_results_

grid = pd.DataFrame(results).loc[:, ('rank_test_MSE', 'mean_test_MSE', 'param_alpha_1', 
                                     'param_alpha_2', 'param_lambda_2')]

grid.sort_values(by='rank_test_MSE')

end = time.time()
m, s = divmod(end - start, 60)
h, m = divmod(m, 60)
print "Finalizó!"
print 'Tiempo:', "%02d:%02d:%02d" % (h, m, s)
